requirements.txt
asgiref==3.5.2
certifi==2022.6.15
distlib==0.3.5
Django==4.0.6
djangorestframework==3.15.1
filelock==3.4.2
platformdirs==2.6.2
six==1.16.0
sqlparse==0.4.2
tzdata==2024.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
